WIP Fix federation directive variables
This is a work-in-progress PR to fix an issue with graphql-go-tools
federation and directive variables. When an entity query contains a
field with a directive that includes a variable, the variable isn't
added to constructed operation.

The first commit adds a failing test that demonstrates the issue. Note
that I had to tweak the testing infrastructure to see the diff in the
failing test with the objects being compared are deeply nested and the
testify assert.Equal function is hitting its configured depth limit.

I verified the test passed when the someSkipCondition isn't present in
the operation arguments or variables object. I think I've added the
variable in the places it'll show up, but it's possible I got it wrong.
Anyway, the exact format of the input doesn't matter as long as the
variable is present!