Fix the json variable parser to properly handle null and non-null fields.
The previous logic was saying that null fields could be omitted from
the input, but not that they could be set explicitly to null.  Now
both are allowed, as per the spec.

For example, if the input is `$a: String`, it is now legal to pass any of:
```
{"a": "some string"}  // always allowed
{}                    // always allowed
{"a": null}           // newly allowed
```

The new tests turned up a few other bugs, most notably misspelling the
jsonschema keyword `items` as `item`.