Debug: planner instrumentation
This PR isn't meant to be landed, but it's super useful for debugging.

The PR instruments the planner so each node type and ref in the
operation being planned is printed before the node is visited. It also
prints when a specific planner will visit a node as well as the upstream
operation after a node is visited.

Also, after the configuration visitor has visited all the nodes, the
planner paths are printed out; these are the paths in the operation each
data source planner is responsible for resolving.