Add normalization to support abstract field federation
This PR adds the three operation normalization passes outlined in #328
to support abstract field federation. The new passes 1) "expand" inline
fragments with interface type conditions into inline fragments with
concrete type conditions, 2) "expand" interface selections to selections
that have inline fragments with the corresponding concrete types, and 3)
eliminate invalid inline fragments as the result of passes (1) and (2).
These passes run after fragment inlining but before inline fragment
merging. In the code I note that we may also want to implement merging
of inline fragment siblings that match.

A lot of tests needed to be updated due to these changes. The REST and
JSON data source tests were updated, but it isn't clear to me if
existing REST and JSON datasources will be compatible with these
changes; they need to return the appropriate "__typename" in responses
(since resolver fields now have OnTypeName conditions). Guidance on this
topic would be appreciated.

I also updated the ValidateByFieldList function to "expand" interface
types used in block and allow conditions to the corresponding concrete
types. This is what I'd expect (if Character "name" isn't allowed, then
Human "name" shouldn't be allowed either), but I could be mistaken about
the exact use of this function.

Depends on #336. (This is a stacked PR; the parent PR is expected to be
merged into master, followed by this PR, instead of merging this PR into
the parent PR directly.)