Fix required fields skipped in output.
Fixes #335.

This PR fixes the case where a required field is present in *some*
fragment but a field in a different fragment requires the field, so the
field is added to the second fragment and entirely skipped in the
response.

The fix required tracking skipped fields by type in addition to path.
Also, it's possible a parent fetch already fetched the required field.
During federation this can cause a fragment selection to be emtpy. I've
updated the deleteInvalidInlineFragments normalization function to also
delete empty inline fragments.

This issue isn't specific to interface federation, but that's when it's
most likely to occur, so I've stacked the PR on top of the abstract
field federation PR, #5.

Depends on #5.