WIP: Fix OnTypeName resolve panic
There's a panic that occurs when resolving an object field and 1) the
object has an associated fetch, and 2) the field being resolved is
skipped because OnTypeName is set. In this case the context
`responseElements` are reset, but they aren't restored when skipping the
field. This results in a panic later when attemping to remove the
existing path elements from the context.

This PR is a work-in-progress because I still need to construct a test
case to demonstrate the issue.